+++ webpack.config.babel.js
@@ -3,6 +3,7 @@ import path from 'path'
 
 import MiniCssExtractPlugin from 'mini-css-extract-plugin'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
+import sass from 'dart-sass'
 
 import config from './package.json'
 
@@ -86,7 +87,10 @@ export default {
           {
             loader: 'sass-loader',
             options: {
-              includePaths: [path.join(__dirname, 'src')],
+              implementation: sass,
+              sassOptions: {
+                includePaths: [path.join(__dirname, 'src')],
+              },
             },
           },
         ],